Moved lager, messaging, gonc, common, etcdsync, plugins and security to pkg
