--- lager/examples/main.go
@@ -1,47 +1,0 @@
-//Copyright 2017 Huawei Technologies Co., Ltd
-//
-//Licensed under the Apache License, Version 2.0 (the "License");
-//you may not use this file except in compliance with the License.
-//You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-//Unless required by applicable law or agreed to in writing, software
-//distributed under the License is distributed on an "AS IS" BASIS,
-//WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//See the License for the specific language governing permissions and
-//limitations under the License.
-package main
-
-import (
-	"fmt"
-
-	"github.com/servicecomb/service-center/lager"
-)
-
-func main() {
-	logger.Init(logger.Config{
-		LoggerLevel:   "DEBUG",
-		LoggerFile:    "",
-		EnableRsyslog: false,
-		LogFormatText: false,
-	})
-
-	logger := logger.NewLogger("example")
-
-	logger.Infof("Hi %s, system is starting up ...", "paas-bot")
-
-	logger.Debug("check-info", lager.Data{
-		"info": "something",
-	})
-
-	err := fmt.Errorf("Oops, error occurred!")
-	logger.Warn("failed-to-do-somthing", err, lager.Data{
-		"info": "something",
-	})
-
-	err = fmt.Errorf("This is an error")
-	logger.Error("failed-to-do-somthing", err)
-
-	logger.Info("shutting-down")
-}
