+++ pkg/lager/core/chug/match_log_entry_test.go
