+++ pkg/lager/core/writer_sink.go
