+++ pkg/lager/core/writer_sink_test.go
