+++ pkg/plugins/loader.go
