+++ server/core/registry/embededetcd/embed_etcd_plugin.go
@@ -23,7 +23,7 @@ import (
 	"github.com/coreos/etcd/etcdserver/etcdserverpb"
 	"github.com/coreos/etcd/lease"
 	"github.com/coreos/etcd/mvcc/mvccpb"
-	"github.com/servicecomb/service-center/common"
+	"github.com/servicecomb/service-center/pkg/common"
 	"github.com/servicecomb/service-center/server/core/registry"
 	"github.com/servicecomb/service-center/util"
 	"github.com/servicecomb/service-center/util/rest"
