+++ util/rest/common.go
@@ -18,8 +18,8 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"github.com/astaxie/beego"
-	"github.com/servicecomb/service-center/common"
-	"github.com/servicecomb/service-center/security"
+	"github.com/servicecomb/service-center/pkg/common"
+	"github.com/servicecomb/service-center/pkg/security"
 	"github.com/servicecomb/service-center/util"
 	"io/ioutil"
 	"time"
